+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `project view --content` flag to display the full project description/overview
+
 ## [0.5.0] - 2026-03-24
 
 ### Added